Initial attempt at dterm cut
If enabled, this code cuts D by a specified percentage durning normal flight, allowing it to smoothly rise back to the normal D value during, and shortly after, rapid gyro moves like flips and rolls, and to a lesser extent during prop wash.
Input is gyro differential data with higher frequencies (above propwash) attenuated with a 50hz biquad 'range' filter).  The absolute value of this is then 'integrated, smoothed and delayed' with a 5hz PT1 lowpass filter.  Using a gain factor we should end up with a control signal that sits around zero during stable normal flight, and peaks at 1.0 during flips and rolls.
When the control signal is zero, the D setting is attenuated by the specified percentage.  A larger percentage cut results in less D during normal flight, but the amount of D available during flips and rolls should be unchanged.  For example, if D is 40, a cut of 30% (a setting of 30 in CLI) will be cut by 12 to 28, a cut of 70% will cut it to 12.